URLencode filter values
E-mail addresses can contain '+' signs, which get converted back into a space ' ' when filtering. This causes the subscriber lookup to fail:

$subscriber = $emailList->subscriber($user->email);

..because the + sign in the email is not there anymore.

A urlencode here prevents confusion in the encoding that happens in http_build_query.

If this causes unexpected behavior with other filters, I'd advise to update the documentation instruct users to urlencode on their end.